backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.4.2
pydantic-settings==2.0.3

# Vector Database and Embeddings
chromadb==0.4.15
sentence-transformers==2.2.2

# Google Gemini Pro
google-generativeai==0.3.0

# LangChain for RAG
langchain==0.0.335
langchain-community==0.0.6

# Data Processing
pandas==2.1.3
numpy==1.24.3

# HTTP and API
httpx==0.25.2
aiofiles==23.2.1

# Environment and Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Logging and Monitoring
structlog==23.2.0 